#87 [feat] 메인 홈 연결 및 검색창 활성화
