+++ app/src/main/AndroidManifest.xml
@@ -57,8 +57,9 @@
         <activity
             android:name=".presentation.ui.mission.create.CreateMissionActivity"
             android:exported="true" />
-
         <activity android:name=".presentation.ui.mission.mymission.view.MissionProgressActivity" />
+        <activity android:name=".presentation.ui.search.view.SearchPostActivity"
+            android:exported="true" />
     </application>
 
 </manifest>
