+++ app/src/main/java/com/dnd_8th_4_android/wery/presentation/ui/search/viewmodel/SearchPostViewModel.kt
@@ -1,0 +1,7 @@
+package com.dnd_8th_4_android.wery.presentation.ui.search.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class SearchPostViewModel : ViewModel() {
+
+}
