+++ app/src/main/res/layout/fragment_home.xml
@@ -24,7 +24,7 @@
             android:paddingVertical="8dp"
             app:layout_constraintTop_toTopOf="parent">
 
-            <EditText
+            <TextView
                 android:id="@+id/et_search"
                 style="@style/TextView.Caption.12.R"
                 android:layout_width="0dp"
@@ -32,8 +32,7 @@
                 android:layout_marginEnd="8dp"
                 android:background="@drawable/shape_gray100_radius_24"
                 android:hint="@string/home_search_hint"
-                android:imeOptions="actionSearch"
-                android:inputType="text"
+                android:gravity="center_vertical"
                 android:paddingHorizontal="38dp"
                 app:layout_constraintEnd_toStartOf="@id/iv_post_bookmark"
                 app:layout_constraintStart_toStartOf="parent"
@@ -46,17 +45,6 @@
                 android:src="@drawable/ic_search"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintStart_toStartOf="@id/et_search"
-                app:layout_constraintTop_toTopOf="parent" />
-
-            <ImageView
-                android:id="@+id/iv_search_close"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginEnd="12dp"
-                android:src="@drawable/ic_search_close"
-                android:visibility="gone"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="@id/et_search"
                 app:layout_constraintTop_toTopOf="parent" />
 
             <ImageView
