+++ app/src/main/res/values/strings.xml
@@ -183,7 +183,7 @@
     <string name="yes_mission_certify">인증하기</string>
     <string name="yes_mission_certify_write">인증글 쓰기</string>
 
-    <!-- SearchActivity -->
+    <!-- SearchPostActivity -->
     <string name="search_post">게시글</string>
     <string name="search_filter">최신순</string>
 </resources>
